requirements.txt
# Core ML
scikit-learn>=1.4.0
umap-learn>=0.5.4
shap>=0.44.0
xgboost>=2.0.0
torch>=2.1.0
optuna>=3.4.0

# Data
pandas>=2.1.0
numpy>=1.25.0
openpyxl>=3.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Experiment Tracking & Model Registry
mlflow>=2.9.0

# Data Versioning
dvc>=3.30.0

# UI
streamlit>=1.29.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Monitoring
prometheus-client>=0.19.0
scipy>=1.11.0
requests>=2.31.0

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
flake8>=6.1.0
black>=23.11.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
joblib>=1.3.0
